modelの処理を修正

+++ ToDoList/Model/ToDoModel.swift
@@ -27,7 +27,7 @@ struct TodoListError: Error {
     private(set) var message: String = ""
 
     enum ErrorType: CaseIterable {
-        case add, update, delete, allDelete
+        case add, update, delete, allDelete, initError
     }
 
     init(type: ErrorType) {
@@ -41,6 +41,8 @@ struct TodoListError: Error {
             message = R.string.message.deleteError()
         case .allDelete:
             message = R.string.message.allDeleteError()
+        case .initError:
+            message = "システムエラーが発生しました"
         }
     }
 
@@ -90,16 +92,15 @@ final class ToDoModel: Object {
     }
 
     /// Realmのインスタンス化
-    private static var initRealm: Result<Realm, Error> {
+    private static var initRealm: Realm? {
         let realm: Realm
         var configuration: Realm.Configuration = Realm.Configuration()
         configuration.schemaVersion = UInt64(1)
-
         do {
             realm = try Realm(configuration: configuration)
-            return .success(realm)
+            return realm
         } catch {
-            return .failure(error)
+            return nil
         }
     }
 
@@ -107,36 +108,30 @@ final class ToDoModel: Object {
     /// - Parameters:
     ///   - addValue: 追加するToDo
     ///   - addError: エラー発生時のクロージャー
-    class func add(addValue: ToDoModel) -> Result<Void, TodoListError> {
-        switch ToDoModel.initRealm {
-        case .success(let realm):
-            let toDoModel = ToDoModel(id: String(allFind()!.count),
-                                      toDoName: addValue.toDoName,
-                                      todoDate: addValue.todoDate,
-                                      toDo: addValue.toDo,
-                                      completionFlag: CompletionFlag.unfinished.rawValue,
-                                      createTime: Format().stringFromDate(date: Date(), addSec: true)
-            )
+    class func add(addValue: ToDoModel) throws {
+        guard let realm = ToDoModel.initRealm else {
+            throw TodoListError(type: .initError)
+        }
+        let toDoModel = ToDoModel(id: String(allFind().count),
+                                  toDoName: addValue.toDoName,
+                                  todoDate: addValue.todoDate,
+                                  toDo: addValue.toDo,
+                                  completionFlag: CompletionFlag.unfinished.rawValue,
+                                  createTime: Format().stringFromDate(date: Date(), addSec: true)
+        )
 
-            do {
-
-                try realm.write {
-                    realm.add(toDoModel)
-                }
-
-                Log.devprint("Todoを作成しました: \(toDoModel)")
-                NotificationManager().addNotification(toDoModel: toDoModel) { result in
-                    NotificationCenter.default.post(name: Notification.Name(R.string.notification.tableReload()), object: nil)
-                    NotificationCenter.default.post(name: Notification.Name(R.string.notification.toast()), object: result)
-                }
-                return .success(Void())
-            } catch {
-                Log.devprint(error.localizedDescription)
-                return .failure(TodoListError(type: .add))
+        do {
+            try realm.write {
+                realm.add(toDoModel)
             }
-        case .failure(let error):
+            Log.devprint("Todoを作成しました: \(toDoModel)")
+            NotificationManager().addNotification(toDoModel: toDoModel) { result in
+                NotificationCenter.default.post(name: Notification.Name(R.string.notification.tableReload()), object: nil)
+                NotificationCenter.default.post(name: Notification.Name(R.string.notification.toast()), object: result)
+            }
+        } catch {
             Log.devprint(error.localizedDescription)
-            return .failure(TodoListError(type: .add))
+            throw TodoListError(type: .add)
         }
     }
 
@@ -144,32 +139,27 @@ final class ToDoModel: Object {
     /// - Parameters:
     ///   - updateValue: 更新するToDo
     ///   - updateError: エラー発生時のクロージャー
-    class func update(updateValue: ToDoModel) -> Result<Void, TodoListError> {
-        switch ToDoModel.initRealm {
-        case .success(let realm):
+    class func update(updateValue: ToDoModel) throws {
+        guard let realm = ToDoModel.initRealm,
+              let toDoModel = ToDoModel.find(todoId: updateValue.id, createTime: updateValue.createTime) else {
+            throw TodoListError(type: .initError)
+        }
 
-            let toDoModel = ToDoModel.find(todoId: updateValue.id, createTime: updateValue.createTime)!
-
-            do {
-                try realm.write {
-                    toDoModel.toDoName = updateValue.toDoName
-                    toDoModel.todoDate = updateValue.todoDate
-                    toDoModel.toDo = updateValue.toDo
-                    toDoModel.completionFlag = CompletionFlag.unfinished.rawValue
-                }
-                Log.devprint("Todoを更新しました: \(toDoModel)")
-                NotificationManager().addNotification(toDoModel: toDoModel) { result in
-                    NotificationCenter.default.post(name: Notification.Name(R.string.notification.tableReload()), object: nil)
-                    NotificationCenter.default.post(name: Notification.Name(R.string.notification.toast()), object: result)
-                }
-                return .success(Void())
-            } catch {
-                Log.devprint(error.localizedDescription)
-                return .failure(TodoListError(type: .update))
+        do {
+            try realm.write {
+                toDoModel.toDoName = updateValue.toDoName
+                toDoModel.todoDate = updateValue.todoDate
+                toDoModel.toDo = updateValue.toDo
+                toDoModel.completionFlag = CompletionFlag.unfinished.rawValue
             }
-        case .failure(let error):
+            Log.devprint("Todoを更新しました: \(toDoModel)")
+            NotificationManager().addNotification(toDoModel: toDoModel) { result in
+                NotificationCenter.default.post(name: Notification.Name(R.string.notification.tableReload()), object: nil)
+                NotificationCenter.default.post(name: Notification.Name(R.string.notification.toast()), object: result)
+            }
+        } catch {
             Log.devprint(error.localizedDescription)
-            return .failure(TodoListError(type: .update))
+            throw TodoListError(type: .update)
         }
     }
 
@@ -177,11 +167,13 @@ final class ToDoModel: Object {
     /// - Parameters:
     ///   - updateTodo: 更新するTodo
     ///   - flag: 変更する値
-    static func updateCompletionFlag(updateTodo: ToDoModel, flag: CompletionFlag) {
-        switch initRealm {
-        case .success(let realm):
-            guard let toDoModel = ToDoModel.find(todoId: updateTodo.id, createTime: updateTodo.createTime) else { return }
-            try? realm.write {
+    static func updateCompletionFlag(updateTodo: ToDoModel, flag: CompletionFlag) throws {
+        guard let realm = ToDoModel.initRealm,
+              let toDoModel = ToDoModel.find(todoId: updateTodo.id, createTime: updateTodo.createTime) else {
+            throw TodoListError(type: .initError)
+        }
+        do {
+            try realm.write {
                 toDoModel.completionFlag = flag.rawValue
             }
             if updateTodo.completionFlag == CompletionFlag.completion.rawValue {
@@ -193,8 +185,9 @@ final class ToDoModel: Object {
                 }
             }
 
-        case .failure(let error):
+        } catch {
             Log.devprint("完了フラグ更新エラー: \(error)")
+            throw TodoListError(type: .update)
         }
     }
 
@@ -203,45 +196,41 @@ final class ToDoModel: Object {
     ///   - todoId: todoId
     ///   - createTime: 作成時間
     class func find(todoId: String, createTime: String?) -> ToDoModel? {
-        switch ToDoModel.initRealm {
-        case .success(let realm):
-            if let _createTime = createTime {
-                let todo = (realm.objects(ToDoModel.self).filter("createTime == '\(String(describing: _createTime))'").first)
-                Log.devprint("Todoを１件表示します: \(String(describing: todo))")
-                return todo
-            } else {
-                let todo = (realm.objects(ToDoModel.self).filter("id == '\(todoId)'").first!)
-                Log.devprint("Todoを１件表示します: \(todo)")
-                return todo
-            }
-        case .failure(let error):
-            Log.devprint(error.localizedDescription)
+        guard let realm = ToDoModel.initRealm else {
+            print(TodoListError(type: .initError).message)
             return nil
+        }
+        if let _createTime = createTime {
+            let todo = (realm.objects(ToDoModel.self).filter("createTime == '\(String(describing: _createTime))'").first)
+            Log.devprint("Todoを１件表示します: \(String(describing: todo))")
+            return todo
+        } else {
+            let todo = (realm.objects(ToDoModel.self).filter("id == '\(todoId)'").first)
+            Log.devprint("Todoを１件表示します: \(todo ?? ToDoModel())")
+            return todo
         }
     }
 
     /// 全件取得
     /// - Parameter vc: 呼び出し元のViewController
     /// - Returns: 取得したTodoを全件返す
-    class func allFind() -> [ToDoModel]? {
-        switch ToDoModel.initRealm {
-        case .success(let realm):
-            var todomodel = [ToDoModel]()
-            let todo = realm.objects(ToDoModel.self)
-            todo.forEach { value in
-                todomodel.append(value)
-            }
-            return todomodel
-        case .failure(let error):
-            Log.devprint(error.localizedDescription)
-            return nil
+    class func allFind() -> [ToDoModel] {
+        guard let realm = ToDoModel.initRealm else {
+            return []
         }
+        var todomodel = [ToDoModel]()
+        let todo = realm.objects(ToDoModel.self)
+        todo.forEach { value in
+            todomodel.append(value)
+        }
+        return todomodel
     }
 
     /// 全件取得
-    class func activeFindToDo(index: CompletionFlag) -> [ToDoModel]? {
-        guard let todomodel = ToDoModel.allFind() else {
-            return nil
+    class func activeFindToDo(index: CompletionFlag) -> [ToDoModel] {
+        let todomodel = ToDoModel.allFind()
+        if todomodel.isEmpty {
+            return []
         }
 
         switch index {
@@ -265,7 +254,7 @@ final class ToDoModel: Object {
             return filterToDo
         }
 
-        return nil
+        return []
     }
 
     /// ToDoの削除
@@ -274,27 +263,24 @@ final class ToDoModel: Object {
     ///   - todoId: TodoId
     ///   - createTime: Todoの作成時間
     ///   - completion: 削除完了後の動作
-    class func delete(_ model: ToDoModel) -> Result<Void, TodoListError> {
-        switch ToDoModel.initRealm {
-        case .success(let realm):
-            guard let toDoModel = ToDoModel.find(todoId: model.id, createTime: model.createTime) else {
-                return .failure(TodoListError(type: .delete))
+    class func delete(_ model: ToDoModel) throws {
+        guard let realm = ToDoModel.initRealm else {
+            throw TodoListError(type: .initError)
+        }
+        guard let toDoModel = ToDoModel.find(todoId: model.id, createTime: model.createTime),
+              let createTime = toDoModel.createTime else {
+            throw TodoListError(type: .delete)
+        }
+        do {
+            Log.devprint("Todoを削除します: \(toDoModel)")
+            try realm.write {
+                realm.delete(toDoModel)
             }
-            NotificationManager().removeNotification([toDoModel.createTime!])
-            do {
-                Log.devprint("Todoを削除します: \(toDoModel)")
-                try realm.write {
-                    realm.delete(toDoModel)
-                }
-                Log.devprint("Todoを削除しました")
-                return .success(Void())
-            } catch {
-                Log.devprint(error.localizedDescription)
-                return .failure(TodoListError(type: .delete))
-            }
-        case .failure(let error):
+            NotificationManager().removeNotification([createTime])
+            Log.devprint("Todoを削除しました")
+        } catch {
             Log.devprint(error.localizedDescription)
-            return .failure(TodoListError(type: .delete))
+            throw TodoListError(type: .delete)
         }
     }
 
@@ -302,23 +288,19 @@ final class ToDoModel: Object {
     /// - Parameters:
     ///   - vc: 呼び出し元のViewController
     ///   - completion: 削除完了後の動作
-    class func allDelete() -> Result<Void, TodoListError> {
-        switch ToDoModel.initRealm {
-        case .success(let realm):
-            do {
-                try realm.write {
-                    realm.deleteAll()
-                    Log.devprint("ToDoを全件削除しました")
-                    NotificationManager().allRemoveNotification()
-                }
-                return .success(Void())
-            } catch {
-                Log.devprint(error.localizedDescription)
-                return .failure(TodoListError(type: .delete))
+    class func allDelete() throws {
+        guard let realm = ToDoModel.initRealm else {
+            throw TodoListError(type: .initError)
+        }
+        do {
+            try realm.write {
+                realm.deleteAll()
+                Log.devprint("ToDoを全件削除しました")
+                NotificationManager().allRemoveNotification()
             }
-        case .failure(let error):
+        } catch {
             Log.devprint(error.localizedDescription)
-            return .failure(TodoListError(type: .delete))
+            throw TodoListError(type: .delete)
         }
     }
